requirements.txt
tabulate==0.8.7
mysql-connector-python==8.0.23